+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - API resource: setting/delete_logo
 - API resource: public/add_public_request - in order to handle public contact forms
 - Documentation on "Settings" page on oxe-web-admin
+- Email notification to admins on new community article (private/add_article)
 
 ### Fixed